disambiguate equality added; all cn float

+++ src/parse.jl
@@ -33,7 +33,7 @@ function parse_cn(node)
     if haskey(node, "type") && elements(node) != EzXML.Node[]
         parse_cn_w_sep(node)
     else
-        Meta.parse(node.content)
+        Float64(Meta.parse(node.content))   # convert to Float64 for CellML compatibility
     end
 end
 
@@ -149,14 +149,14 @@ parse a <lambda> node
 ```xml
 <lambda>
   <bvar> x1 </bvar><bvar> xn </bvar>
-   expression-in-x1-xn 
+   expression-in-x1-xn
 </lambda>
 ```
 """
 function parse_lambda(node)
     es = elements(node)
     vars = findall("//x:bvar | //bvar", node, ["x" => MathML.mathml_ns])
-    # vars2 = findall("//bvar", node) # works in tests 
+    # vars2 = findall("//bvar", node) # works in tests
     # vars = union(vars, vars2) # FIX
 
     args = first.(parse_bvar.(vars))
